deletedAt fix

+++ lib/base/repositories/firestore_repository.dart
@@ -16,26 +16,63 @@ abstract class FirestoreRepository<T extends Entity> extends Repository<T> {
 
   @override
   void add(T model) {
-    collectionRef.doc(model.id).set(serializers.serialize(model));
+    final doc = collectionRef.doc(model.id);
+    final batch = _firestore.batch();
+
+    batch.set(doc, serializers.serialize(model));
+
+    batch.set(
+      doc,
+      {
+        "deletedAt": null,
+      },
+      SetOptions(merge: true),
+    );
+
+    batch.commit();
   }
 
   @override
   void update(T model) {
-    collectionRef.doc(model.id).set(serializers.serialize(model));
+    collectionRef.doc(model.id).set(
+          serializers.serialize(
+            model.rebuild(
+              (b) => b..updatedAt = DateTime.now(),
+            ),
+          ),
+        );
   }
 
   @override
-  void remove(T model) {
-    collectionRef.doc(model.id).delete();
+  void delete(
+    T model, {
+    bool deleteCompleteley = false,
+  }) {
+    if (deleteCompleteley) {
+      collectionRef.doc(model.id).delete();
+    } else {
+      collectionRef.doc(model.id).set(
+            model.rebuild(
+              (b) => b..deletedAt = DateTime.now(),
+            ),
+          );
+    }
   }
 
-  @override
-  Stream<BuiltList<T>> get stream => collectionRef
-      .snapshots()
+  Stream<BuiltList<T>> _deserialize(Stream<QuerySnapshot<Object?>> data) => data
       .map((it) => it.docs)
       .map((it) => it.map((e) => e.data()))
       .map((it) => it.map((e) => deserialize<T>(e)))
       .map((event) => event.whereType<T>().toBuiltList());
+
+  @override
+  Stream<BuiltList<T>> get stream => _deserialize(
+        collectionRef.where("deletedAt", isNull: true).snapshots(),
+      );
+
+  Stream<BuiltList<T>> get streamDeleted => _deserialize(
+        collectionRef.where("deletedAt", isNull: false).snapshots(),
+      );
 
   @override
   // TODO: implement data
